feat: [DPE-5769] - build snap from launchpad source on 24 04 (#3)
This pull request introduces several significant changes to the `etcd`
configuration and setup process. The most important changes include the
addition of a configuration file, a license file, an install hook
script, and a Snapcraft configuration file.

### Configuration and Setup Changes:

* Added a Snapcraft configuration file `snap/snapcraft.yaml` to define
the Snap package for `etcd`, including details such as name, version,
summary, description, confinement, and parts for building the Snap.

* Added a default `etcd` configuration file to `config/etcd.conf.yml`,
which includes settings for data directories, snapshot counts, peer and
client URLs, security settings, and logging configurations.

* Added an install hook script `snap/hooks/install` to copy the `etcd`
configuration file to the Snap's data directory during installation.

* Added an Apache License 2.0 file to `licenses/LICENSE-snap` to define
the terms and conditions for using, reproducing, and distributing the
software.